refactor(frontend): landing componend separate in landing folder ss-53
