+++ apps/frontend/src/pages/landing/landing.tsx
@@ -1,4 +1,4 @@
-import { HeroSection } from "./libs/components/hero-section/hero-section.js";
+import { HeroSection } from "../libs/components/hero-section/hero-section.js";
 import styles from "./styles.module.css";
 
 const Landing = (): React.JSX.Element => {
